Started making the map generating functions.
Created a new class named 'RectangularRoom', with several functions embedded and another function named 'generate_dungeon' which calls on' RectangularRoom' and 'GameMap' to carve out rooms from the walls in the map.

Made changes to 'main.py' so it will call on the 'generate_dungeon' function function when making the map, instead of GameMap directly.