Procedural generation
In previous iterations I was simply testing the room creation functions using manually defined coordinates. Here, I've actually reconstructed the 'generate_dungeon' function, that now creates a random number of rooms between 1 and 'max_rooms', complete with random coordinates and sizes.

As it is, it also calls on the 'intersects' function as well, and if a room it generates would overlap another then it throws it out and tries again.